improvements
Updated so that if enter pressed, the book autocompletes to the first suggestion. Also, suggestions updated to include any occurrence of substring if first characters don't match, and full list available when a complete book name is entered, so that repeated searches are made quicker.
Verse entry updated to allow comma separated verses, with new line and verse number added when a discontinuity occurs.